app/requirements.txt
Django~=4.2
djangorestframework~=3.14.0
drf-spectacular~=0.27.2
psycopg2-binary~=2.9.6
djangorestframework-simplejwt~=5.3.1
loguru~=0.7.2
